// src/types/topics.ts
export enum KafkaTopics {
  START_TRANSACTIONS_CREDIT_CARD = 'start-transactions-credit-card',
  FMS_COMPENSATION = 'fmsCompensation',
  FMS_MOVEMENTS = 'fmsMovements',
  FMS_MOVEMENTS_REPLAY = 'fmsMovementsReplay',
  END_TRANSACTIONS_CREDIT_CARD = 'end-transactions-credit-card',
  FMS_SUCCESS = 'fmsSuccess',
}